Refactor: IntroViewController 제거 후 화면 전환을 SceneDelegate에서 처리
