+++ PRTracker/Resource/Base.lproj/Main.storyboard
@@ -10,10 +10,10 @@
         <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
     </dependencies>
     <scenes>
-        <!--Intro View Controller-->
+        <!--View Controller-->
         <scene sceneID="gOo-ch-unv">
             <objects>
-                <viewController id="HEn-lG-uUf" customClass="IntroViewController" customModule="PRTracker" customModuleProvider="target" sceneMemberID="viewController">
+                <viewController id="HEn-lG-uUf" sceneMemberID="viewController">
                     <view key="view" contentMode="scaleToFill" id="oJ9-lQ-nZP">
                         <rect key="frame" x="0.0" y="0.0" width="414" height="896"/>
                         <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
@@ -386,7 +386,7 @@
         </scene>
     </scenes>
     <inferredMetricsTieBreakers>
-        <segue reference="toQ-Cc-Ot1"/>
+        <segue reference="Gxo-dv-5ho"/>
     </inferredMetricsTieBreakers>
     <resources>
         <namedColor name="CustomBlack">
